+++ SpatialGDK/Source/SpatialGDKEditorToolbar/Public/SpatialGDKSimulatedPlayerDeployment.h
@@ -43,8 +43,6 @@ private:
 	/** Pointer to the SpatialGDK editor */
 	TWeakPtr<FSpatialGDKEditor> SpatialGDKEditorPtr;
 
-	TFuture<bool> AttemptSpatialAuthResult;
-
 	/** Delegate to commit project name */
 	void OnProjectNameCommitted(const FText& InText, ETextCommit::Type InCommitType);
 
@@ -114,7 +112,6 @@ private:
 	void OnBuildConfigurationPicked(FString Configuration);
 
 	FReply OnBuildAndUploadClicked();
-	bool CanBuildAndUpload() const;
 
 	ECheckBoxState ForceAssemblyOverwrite() const;
 	void OnCheckedForceAssemblyOverwrite(ECheckBoxState NewCheckedState);
@@ -123,9 +120,6 @@ private:
 	ECheckBoxState IsUsingGDKPinnedRuntimeVersion() const;
 	bool IsUsingCustomRuntimeVersion() const;
 	FText GetSpatialOSRuntimeVersionToUseText() const;
-
-	/** Delegate to determine the 'Launch Deployment' button enabled state */
-	bool IsDeploymentConfigurationValid() const;
 
 	ECheckBoxState IsBuildClientWorkerEnabled() const;
 	void OnCheckedBuildClientWorker(ECheckBoxState NewCheckedState);
@@ -137,8 +131,6 @@ private:
 	void OnCheckedGenerateSnapshot(ECheckBoxState NewCheckedState);
 
 	void OnBuildSuccess();
-
-	bool CanLaunchDeployment() const;
 
 	FReply OnOpenCloudDeploymentPageClicked();
 	bool CanOpenCloudDeploymentPage() const;